backend/standalone/requirements.txt
fastapi
python-socketio
uvicorn[standard]
numpy
pandas
